**Changes:**
- Use float4 to force aligned reads of 128bit to increase memory bandwidth
- Increase block dimension to 32x4 to submit enough threads per SM to avoid stalls


**Performance**
Using llama-bench I measured the end-to-end speedup 
Device 0: NVIDIA RTX 6000 Ada Generation, compute capability 8.9, VMM: yes</p>

BASLINE
model | size | params | backend | ngl | test | t/s
-- | -- | -- | -- | -- | -- | --
llama 7B Q4_K - Medium | 3.80 GiB | 6.74 B | CUDA | 100 | pp128 | 6476.13 ± 66.28
llama 7B Q4_K - Medium | 3.80 GiB | 6.74 B | CUDA | 100 | tg128 | 140.66 ± 0.65

FLOAT4
model | size | params | backend | ngl | test | t/s
-- | -- | -- | -- | -- | -- | --
llama 7B Q4_K - Medium | 3.80 GiB | 6.74 B | CUDA | 100 | pp128 | 6479.48 ± 107.08
llama 7B Q4_K - Medium | 3.80 GiB | 6.74 B | CUDA | 100 | tg128 | 142.87 ± 0.39


128 Threads per Block
model | size | params | backend | ngl | test | t/s
-- | -- | -- | -- | -- | -- | --
llama 7B Q4_K - Medium | 3.80 GiB | 6.74 B | CUDA | 100 | pp128 | 6489.03 ± 118.45
llama 7B Q4_K - Medium | 3.80 GiB | 6.74 B | CUDA | 100 | tg128 | 145.76 ± 0.88

FLOAT4 + 128 Threads per Block
model | size | params | backend | ngl | test | t/s
-- | -- | -- | -- | -- | -- | --
llama 7B Q4_K - Medium | 3.80 GiB | 6.74 B | CUDA | 100 | pp128 | 6522.17 ± 98.75
llama 7B Q4_K - Medium | 3.80 GiB | 6.74 B | CUDA | 100 | tg128 | 149.49 ± 0.28


- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [X] Low
  - [ ] Medium
  - [ ] High