@slaren 2024-06-26
With `LLAMA_CUDA_FORCE_DMMV` I see a significant improvement from these changes:

| GPU         | Model           | Test   |   t/s master |   t/s q4_k_performance |   Speedup |
|:------------|:----------------|:-------|-------------:|-----------------------:|----------:|
| RTX 3090 Ti | llama 7B Q4_K_M | tg128  |       104.36 |                 127.25 |      1.22 |


However, without `LLAMA_CUDA_FORCE_DMMV` mmvq is used instead, which is significantly faster:

| model                          |       size |     params | backend    | ngl |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | ---------------: |
| llama 7B Q4_K - Medium         |   3.80 GiB |     6.74 B | CUDA       |  99 |         tg128 |    146.10 ± 1.12 |

For this format I think mmvq should be automatically used for compute capability >= 6.1.