@ggerganov 2024-06-26
Similar results for RTX 2060:

```bash
LLAMA_CUDA_FORCE_DMMV=1 LLAMA_CUDA=1 ./scripts/compare-commits.sh master pr/8136 -m ./models/llama-7b-v2/ggml-model-q4_k.gguf -p 0 -ngl 99
```

| GPU            | Model           | Test   |   t/s master |   t/s pr/8136 |   Speedup |
|:---------------|:----------------|:-------|-------------:|--------------:|----------:|
| RTX 2060 SUPER | llama 7B Q4_K_M | tg128  |        29.47 |         40.27 |      1.37 |

```bash
LLAMA_CUDA=1 make -j && ./llama-bench -m ./models/llama-7b-v2/ggml-model-q4_k.gguf -p 0 -ngl 99
```

ggml_cuda_init: GGML_CUDA_FORCE_MMQ:    no
ggml_cuda_init: GGML_CUDA_FORCE_CUBLAS: no
ggml_cuda_init: found 1 CUDA devices:
  Device 0: NVIDIA GeForce RTX 2060 SUPER, compute capability 7.5, VMM: yes
| model                          |       size |     params | backend    | ngl |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | ---------------: |
| llama 7B Q4_K - Medium         |   3.80 GiB |     6.74 B | CUDA       |  99 |         tg128 |     69.08 ± 0.24 |

build: 50732c06 (3236)