@JohannesGaessler 2024-06-26
>Use float4 to force aligned reads of 128bit to increase memory bandwidth

I don't think that's the reason. `__builtin_assume_aligned` does not achieve any speedup. And unlike with e.g. `half2` there is no hardware support for `float4`; the compiler will simply convert it to float instructions. I think the reason for the performance difference is rather that with the code on master the compiler fails to group or schedule the memory accesses to `y1` and `y2` in the optimal way.

In any case, I can confirm that the performance increases with this change and that most of the difference comes from the `float4` change.