@JohannesGaessler 2024-06-26
>And unlike with e.g. half2 there is no hardware support for float4; the compiler will simply convert it to float instructions.

Unless I am completely misremembering the CUDA documentation definitely said that the `float4` datatype effectively is just 4 floats. However, looking at the PTX code it seems that there is a dedicated instruction for loading a chunk of 16 bytes:

```
1031	ld.global.nc.v4.f32 	{%f9, %f10, %f11, %f12}, [%rd31+-512];
1032	
1033	ld.global.nc.v4.f32 	{%f17, %f18, %f19, %f20}, [%rd31+-384];
1034	
1035	ld.global.nc.v4.f32 	{%f25, %f26, %f27, %f28}, [%rd31];
1036	
1037	ld.global.nc.v4.f32 	{%f33, %f34, %f35, %f36}, [%rd31+128];
```

Assuming this can also be applied to other kernels this is definitely a very good find.