@contentis 2024-06-27
@JohannesGaessler float4 issues a single instruction to read 128bit. In theory, the compiler should be capable of understanding this from the original code (w.o. having to use "unsafe" methods as reinterpret cast), but the compiler often is imperfect. 

@slaren My plan was to look at `dequantize_mul_mat_vec` (Q4_0) next and see if I can find opportunities for optimization. 
Would you say mmvq is used more frequently? Trying to understand what is the most common format.