@slaren 2024-06-27
> @slaren My plan was to look at `dequantize_mul_mat_vec` (Q4_0) next and see if I can find opportunities for optimization.
> Would you say mmvq is used more frequently? Trying to understand what is the most common format.

Yes, mmvq is used more frequently. At this point, the dmmv kernels are only with GPUs that don't support dp4a, cc < 6.1.